Add save/load session funcionality (#530105)
Allows the user to save the current upload session for later.
This preserves any titles, descriptions, tags, photosets, groups,
permissions, and visibility settings for any files in the upload
window.

This file can be loaded as well.  There are a few edge cases
with photoset data if the currently logged in user does not
match the user that performed the export.  In those cases
photoset data is dropped.

Also adds a "Save" option to the quit confirm dialogs to
export the pending upload set.

Signed-off-by: Germán Póo-Caamaño <gpoo@gnome.org>