Allow the creation of a new photoset in an upload session
Fix #507232.

This patch adds another default entry in the "Add to Set" combobox
to create a photoset w/the upload.  By default, the photoset will
be named, "new photoset (%m-%d-%y)" but can be renamed via the
"rename" button that appears next to it when this option is selected.

If the upload is interrupted, or if the upload session is saved,
the new photoset's name will be preserved on load.  The newly created
photoset behaves the same way as an already existing set, in that all
photos must be specified to the new set in order to be included.
Only one new photoset can be created per upload.

Signed-off-by: Germán Póo-Caamaño <gpoo@gnome.org>