Eliminate race conditions in PyUnique.
A race condition was present in the pyunique code, where several
applications could be started at the same time, each trying to register
the DBus objects at the same time. This is solved by requiring the
application to have registered a special DBus name ("lock") before they
can move on. Credits to Mathias Hasselmann for inspiration and DBus
magic hints.